乱数登録処理をフロントに追加
